feat(community): split MyInfoCardList into two components
